+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-﻿# 0.6.0
+﻿# 0.6.2
+<sup>Released: 2017/8/20</sup>
+
+- Change method trace string to have full namespace, method name and parameters.
+
+# 0.6.1
+<sup>Released: 2017/8/20</sup>
+
+- Fixed a bug where .NET Core Service extensions were excluded from the library.
+
+# 0.6.0
 
 ## Features
 <sup>Released: 2017/8/20</sup>